Add Conservation calculations to temperature and humidity data

Appends columns for conservation-risks: mould risk, preservation indices, equilibrium moisture, and moisture content for wood to a dataframe with temperature and relative humidity columns.

Usage

add_conservation_calcs(mydata, Temp = "Temp", RH = "RH", ...)

Arguments

mydata

A dataframe containing temperature and relative humidity data.

Temp

Character string name of the temperature column (default "Temp").

RH

Character string name of the relative humidity column (default "RH").

...

Additional parameters passed to humidity calculation functions.

Value

Dataframe augmented with conservation variables:

- Mould_LIM: Mould risk threshold humidity (numeric, from Zeng equation). - Mould_rate: Mould growth rate index (from Zeng equation, labelled output). - Mould_index: Mould risk index (continuous scale, VTT model). - PreservationIndex: Preservation Index for collection longevity. - Lifetime: Lifetime Multiplier for object material degradation risk. - EMC_wood: Wood equilibrium moisture content (

Examples

# mydata file
filepath <- data_file_path("mydata.xlsx")
mydata <- readxl::read_excel(filepath, sheet = "mydata", n_max = 5)

mydata |> add_conservation_calcs() |> dplyr::glimpse()
#> Rows: 5
#> Columns: 11
#> $ Site              <chr> "London", "London", "London", "London", "London"
#> $ Sensor            <chr> "Room 1", "Room 1", "Room 1", "Room 1", "Room 1"
#> $ Date              <dttm> 2024-01-01 00:00:00, 2024-01-01 00:15:00, 2024-01-01…
#> $ Temp              <dbl> 21.8, 21.8, 21.8, 21.7, 21.7
#> $ RH                <dbl> 36.8, 36.7, 36.6, 36.6, 36.5
#> $ Mould_LIM         <dbl> 75.11542, 75.11542, 75.11542, 75.14014, 75.14014
#> $ Mould_rate        <dbl> 0, 0, 0, 0, 0
#> $ Mould_index       <dbl> 0, 0, 0, 0, 0
#> $ PreservationIndex <dbl> 45.25849, 45.38181, 45.50580, 46.07769, 46.20393
#> $ Lifetime          <dbl> 1.107855, 1.108860, 1.109869, 1.109854, 1.110866
#> $ EMC_wood          <dbl> 7.201471, 7.186361, 7.171247, 7.173308, 7.158186