Calculate Mould Growth Probability

This function calculates the probability of mould growth based on temperature and relative humidity data.

Usage

calcMould(mydata, Temp, RH)

Arguments

mydata

A data frame containing temperature and relative humidity data.

Temp

The name of the column in mydata containing temperature data. Default is "Temp".

RH

The name of the column in mydata containing relative humidity data. Default is "RH".

Value

A data frame with additional columns:

time_diff

Time difference between consecutive measurements

mould_prob

Probability of mould growth for each time step

mould

Cumulative mould growth probability

Examples

head(mydata) |> calcMould(Temp = "Temp", RH = "RH")
#> # A tibble: 6 × 11
#>   Site   Sensor Date                 Temp    RH TempMould RHMould days_mould
#>   <chr>  <chr>  <dttm>              <dbl> <dbl>     <dbl>   <dbl>      <dbl>
#> 1 London Room 1 2024-01-01 00:00:00  21.8  36.8        22      37         NA
#> 2 London Room 1 2024-01-01 00:15:00  21.8  36.7        22      37         NA
#> 3 London Room 1 2024-01-01 00:29:59  21.8  36.6        22      37         NA
#> 4 London Room 1 2024-01-01 00:44:59  21.7  36.6        22      37         NA
#> 5 London Room 1 2024-01-01 00:59:59  21.7  36.5        22      36         NA
#> 6 London Room 1 2024-01-01 01:14:59  21.7  36.2        22      36         NA
#> # ℹ 3 more variables: time_diff <drtn>, mould_prob <dbl>, mould <dbl>