Calculate Mould Growth Probability Using Various Models

This function calculates the probability or critical conditions for mould growth based on temperature and relative humidity, using one of four different models.

Usage

calcMould_models(Temp, RH, model = c("Days", "Sedlbauer", "VIR", "Isaksson"))

Arguments

Temp

Numeric. Temperature (°Celsius)

RH

Numeric. Relative Humidity (0-100%)

model

Character. The model to use for calculation. Options are "Days", "Sedlbauer", "VIR", or "Isaksson".

Value

Numeric. The result depends on the chosen model:

Table: returns days to mould

Sedlbauer and VIR: Critical RH for mould growth from temperature

Isaksson: Daily dose (1/t(ms)) / D(rel/day)

Details

The results are very different so take care in interpretation.

Examples

calcMould_models(20, 80, "Days")
#> [1] 14
calcMould_models(20, 80, "Sedlbauer")
#> [1] 74.4
calcMould_models(25, 70, "VIR")
#> [1] 79.125
calcMould_models(22, 85, "Isaksson")
#> [1] 0.01162899

head(mydata) |>
  dplyr::mutate(
    # Days_mould = calcMould_models(Temp, RH, "Days"),
    RH_crit_Sedlbauer = calcMould_models(Temp, RH, "Sedlbauer"),
    RH_crit_VIR = calcMould_models(Temp, RH, "VIR"),
    Mould_daily_dose = calcMould_models(Temp, RH, "Isaksson")
    )
#> # A tibble: 6 × 8
#>   Site   Sensor Date                 Temp    RH RH_crit_Sedlbauer RH_crit_VIR
#>   <chr>  <chr>  <dttm>              <dbl> <dbl>             <dbl>       <dbl>
#> 1 London Room 1 2024-01-01 00:00:00  21.8  36.8              73.5        79.0
#> 2 London Room 1 2024-01-01 00:15:00  21.8  36.7              73.5        79.0
#> 3 London Room 1 2024-01-01 00:29:59  21.8  36.6              73.5        79.0
#> 4 London Room 1 2024-01-01 00:44:59  21.7  36.6              73.5        79.0
#> 5 London Room 1 2024-01-01 00:59:59  21.7  36.5              73.5        79.0
#> 6 London Room 1 2024-01-01 01:14:59  21.7  36.2              73.5        79.0
#> # ℹ 1 more variable: Mould_daily_dose <dbl>