Create a Psychrometric Chart

Usage

graph_psychrometric(
  mydata,
  Temp,
  RH,
  LowT = 16,
  HighT = 25,
  LowRH = 40,
  HighRH = 60,
  Temp_range = c(0, 40),
  y_func = calcMR,
  ...
)

Arguments

mydata

A data frame containing temperature and relative humidity data.

Temp

Column name in mydata for temperature values.

RH

Column name in mydata for relative humidity values.

LowT

Numeric value for lower temperature limit of the target range. Default is 16°C.

HighT

Numeric value for upper temperature limit of the target range. Default is 25°C.

LowRH

Numeric value for lower relative humidity limit of the target range. Default is 40

HighRHNumeric value for upper relative humidity limit of the target range. Default is 60Temp_rangeNumeric vector of length 2 specifying the overall temperature range for the chart. Default is c(0, 40).y_funcFunction to calculate y-axis values. Default is calcMR (mixing ratio)....Additional arguments passed to y_func. A ggplot object representing the psychrometric chart. This function generates a psychrometric chart based on input temperature and relative humidity data.Various psychrometric functions can be used for the y-axis.

calcHR: Humidity Ratio (g/kg)

calcMR: Mixing Ratio (g/kg)

calcAH: Absolute Humidity (g/m³)

calcSH: Specific Humidity (g/kg)

calcAD: Air Density (kg/m³)

calcDP: Dew Point (°C)

calcEnthalpy: Enthalpy (kJ/kg)

calcPws: Saturation vapor pressure (hPa)

calcPw: Water Vapour Pressure (hPa)

calcPI: Preservation Index

calcIPI: Years to Degradation of reference material

calcLM: Lifetime

# Basic usage with default settings graph_psychrometric(head(mydata, 100), Temp, RH)# Custom temperature and humidity ranges graph_psychrometric(head(mydata, 100), Temp, RH, LowT = 8, HighT = 28, LowRH = 30, HighRH = 70)# Using a different psychrometric function (e.g., Absolute Humidity) graph_psychrometric(head(mydata, 100), Temp, RH, y_func = calcAH)# Adjusting the overall temperature range of the chart graph_psychrometric(head(mydata, 100), Temp, RH, Temp_range = c(12, 30))