Run ConSciR Psychrometric and Graphing Application

Shiny application to upload data to a psychrometric chart. Also includes graphs for temperature and humidity - line plot with limits shaded and a bivariate plot with box showing limits.

CSV or Excel formatted data with "Temp" and "RH" columns can be uploaded to the application.

Use the sliders and functions to set the limits and parameters to be used.

Functions available:

calcHR: Humidity Ratio (g/kg)

calcMR: Mixing Ratio (g/kg)

calcAH: Absolute Humidity (g/m^3)

calcSH: Specific Humidity (g/kg)

calcAD: Air Density (kg/m^3)

calcDP: Dew Point (°C)

calcEnthalpy: Enthalpy (kJ/kg)

calcPws: Saturation vapor pressure (hPa)

calcPw: Water Vapour Pressure (hPa)

calcPI: Preservation Index

calcLM: Lifetime

calcEMC_wood: Equilibrium Moisture Content (wood)

Usage

run_Psychrometric_app()

Value

Shiny object

Examples

if(interactive()) {
    run_Psychrometric_app()
}