Server module to read TRH data from an Excel or CSV file into Shiny application

Server module to create a user interface for uploading Excel (.xlsx or .xls) or CSV files into a Shiny application. It allows users to select specific columns for Date, Temperature, and Relative Humidity from the uploaded data.

Usage

shiny_DataUploaderServer(
  id,
  mydata = NULL,
  Date = "Date",
  Temp_col = "Temp",
  RH_col = "RH",
  Sensor_col = "Sensor",
  Site_col = "Site"
)

Arguments

id

The module ID

mydata

A data frame containing initial data (if any). Default is NULL.

Date

A string specifying the default name of the Date column. Default is "Date".

Temp_col

A string specifying the default name of the Temperature column. Default is "Temp".

RH_col

A string specifying the default name of the Relative Humidity column. Default is "RH".

Sensor_col

A string specifying the default name of the Sensor column. Default is "Sensor".

Site_col

A string specifying the default name of the Site column. Default is "Site".

Value

A list containing two reactive elements:

data

A reactive value containing the uploaded data

selected_columns

A reactive expression that returns a list of the selected column names for Date, Temperature, and RH

Examples

if (FALSE) { # \dontrun{

# In ui.R
uiOutput("file_upload")
uiOutput("column_Date")
uiOutput("column_Temp")
uiOutput("column_RH")

# In server.R
data_upload <- shiny_DataUploader()

# Access the uploaded data
uploaded_data <- data_upload$data()

# Access the selected column names
selected_cols <- data_upload$selected_columns()
} # }